Create rabbitmq_max_channels_per_conn metric
Currently, we alarm if connection['channels'] > 1.  This change creates
a new metric called rabbitmq_max_channels_per_conn which we can alarm
against using a MaaS alarm.  Additionally, this change updates
CONNECTIONS_URL so we only return the data from /api/connections that
we need (channels).

TODO: We will need to create an alarm on rpc-extras which ideally
      alarms if rabbitmq_max_channels_per_conn > 10.  Running a
      dev cluster periodically sees rabbitmq_max_channels_per_conn > 1
      and 10 is an 'arbitrary but likely reasonable limit' which we can
      use to prevent channel ballooning.

Closes issue #188