synchronize kilo branch with all plugins in rpc-openstack
We have a requirement to move the plugins back out to rpc-maas, in
order that other groups can consume (and contribute to) the plugins
more easily.

This commit synchronises the plugins here with those that exist inside
rpc-openstack at this time.

Note - there is a funky way of preserving all history of commits to
these files when they were in rpc-openstack (using git filter-branch).
Sadly, there is very little history to preserve since we lost it in
the commit which moved all plugins to the plugins directory (7a3cd17)
because 'git mv' was not used from what I can tell. This means that I
can't cleanly apply that history here, and so a big commit including
all the changes is the only way to go. Sorry.