Check json_resp before using max() Check the JSON response body for valid data when requesting channels.
After discussions with @git-harry and thinking the error over again, I've
decided that it is not technically an error to not have channels configured.
It is an error to receive an empty response when checking the overview and
nodes API endpoints. I will fix those in a later bug report.

Closes-Bug: #442