Make all a group only
All is a group and a host in the test inventory currently.
This can cause messages like the following in the gates:

 [WARNING]: Found both group and host with same name: all

This makes the "all" name a group, and use localhost as
the hostname.

If we want to keep using groups instead of hosts, then
we should replace host_vars folder to group_vars, and
use :children in the inventory.