Test RPC Maas on Newton
We are still delivering Newton cycle right now, and the images
used in CI are Trusty. This makes sense to checkout the pip
install role with Newton only.

We should increase our test matrix by having multiple branches,
multiple a-r-r, and multiple OSes later.

This temporarily reverses the branch for pip install to Newton,
in order to avoid the absence of a vars/ubuntu14.04 file when
(like currently) running on master.