Restructured the RPC-MAAS role
This change restructures the rpc-maas tasks so that there's reduced
duplication and a lot more clarity regarding the tasks being executed.
The task files are all now prefixed with "maas_" and grouped based on
their various concerns.

This also bring with it the first set of dynamically generated checks
which is starting with the basic CDM checks. These checks have been
moved into the vars/main.yml file. This removes the template sprawl we
have and allows checks to be generated with a common core.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>