Provide more flexible and reliable apt repo config
This patch adds the ability to use a different apt
repository for the maas packages using an alternative
key set and URL configuration. This is to facilitate
the ability to artifact the MaaS apt repo and
configure RPC-O to use the artifacted repo instead of
the upstream source.

Retries are added to the apt key addition to improve
reliability and the conditional for the repo/key
additions are removed as the values are defaulted
and are therefore always defined.