Fix RabbitMQ status plugin for swift-only control planes
This commit fixes a `ValueError: max() arg is an empty sequence` error
encountered in some environments.  This error is triggered when the plugin
runs while there are no connections reported by RabbitMQ, which is an expected
scenario in a swift-only or keystone-only environment.

The change effectively adds a '0' default to max()'s arguments.

Connects: #253