Implement Role Breakdown
This change implements the role break down approved in issue 255. This
change allows us to move away from the single role into a complete set
of playbooks which provides additional flexibility and speed when
deploying RPC-MaaS. This change also allows us to begin exploring ways
to add additional metric collection capabilities to the stack without
having to deal with the technical debt accumulated in the legacy role.

High-Level Overview:
  The RPC-MaaS monitoring system is comprised of a set of playbooks
allowing the deployer to either deploy everything using a single command
or to choose which components are required to monitor a stack. This
means that the monitoring components have been separated out from one
another allowing folks to deploy and monitor stacks like stand alone
swift, cinder, ceph, etc.

Upgrade-ability:
  Care has been taken to ensure that all of the checks deployed by this
change are backwards compatible with all previously known versions of an
RPC-MaaS deployment. This means deployers should be able to immediately
transition to this new deployment methodology without impacting existing
monitoring capabilities.

Usage:
  While these playbooks can be used with an OpenStack-Ansible deployment
they can also be used with a stand-alone inventory. See the
`test/inventory` file for more on how it can be used. As stated previously,
deployers are no longer required to have complete OpenStack ready
inventory, they can now pick and choose the components required without
having to worry about missing items and unknown handling of checks.

Testing:
  These playbooks have also been built to be 100% idempotent. At present
the testing is being done on a complete OpenStack cloud however in
future PRs we'll break down testing to individual components and ensure
that all of the playbooks are testable on their own.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>