Remove OSA dependency
When installing the swift checks the swift client is needed. Prior to
this commit the tasks depended on OSA and the OSA release to upload
objects into swift to create the needed checks. This change removes
the artificial OSA dependency in favor of using the client versions
installed into the maas-venv.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>