Add support for MaaS Poller
This change adds support for rackspace-monitoring-poller, similarly to
how rackspace-monitoring-agent is handled. The appropriate poller configs
are put into place, new private versions of lb_api_check_* templates are
added for the poller along with ping and ssh checks, and the poller is
started and restarted when necessary throughout the process on both
Ubuntu 14.04 and 16.04.

In order to have the poller installed and setup, the
maas_private_monitoring_enabled variable needs to be set to True,
and a value for maas_private_monitoring_zone must be set.

Fixes rcbops/rpc-openstack#2184