resolve issues with extra quoted passwords
Sometimes the password within the openrc file will contain extra quotes
around the variable name when it's read in by the auth_details parser
within `maas_common.py`. If this happens it causes an auth failure when
running the maas checks. This change improves handling with V3 auth and
ensures the password variable does not contain extra quotes.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>