Exclude notifications queues under condition
The notifications queues can only be execluded from monitoring
if one or more consumer is present on the queue.
If the queues are excluded from monitoring but no consumer is present,
it will lead to instabilities of RabbitMQ, ultimately stopping
operations due to system limitations (memory/disk).
The notification_driver need to be properly set to noop if the
messages generation for the notification queues should be suppressed.
Additionally the new versioned_notifications queue is included in this
check.

Closes-Bug: #286