UG-634 Add vars file for verification dependencies
Deploying rpc-maas on a multi-node fails on the verification stage,
because we're not explicitly installing the futures module on all
hosts.  It appears to be getting pulled in implicitly on some hosts
via other modules and therefore isn't failing across the board.

This commit simply creates a new vars file for the verification stage,
and updates maas-verify.yml to install this new pip package set.

Lastly, we remove futures from test-requirements.txt since it is now
installed via the maas-verify.yml playbook.