Add additional constraints and vendor URI
The constraints needed in the test repo need to be updated to ensure
pyrax works 100% of the time.

The URI module in early versions of Ansible is broken. To fix this the
URI module has been vendored to ensure that it will always work for our
limited needs. To allow this module to work its using the deprecated
syntax such that it will not conflict with the actual URI module as
provided by Ansible.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>