LA-367 Install maas_telegraf_format.py dependency
The script `maas_telegraf_format.py` requires `PyYAML` to be installed.
On an AIO this dependency is met because the AIO host is also the
deployment node. When there are multiple hosts, such as a multi-node
AIO, `PyYAML` is missing.

This commit installs `PyYAML` into the `maas_venv` and uses the Python
binary in that virtual env to run the script.