LA-367 Install virtualenv prior to telegraf
The package `virtualenv` needs installing prior to deploying
`maas_telegraf_format.py`. In general, due to the ordering of MaaS
playbooks in rpc-openstack, it is already installed. However, currently
rpc-gating configures MaaS separately and so `virtualenv` is not
installed when `maas-tigkstack-telegraf.yml` runs.

In addition, two `delegate_to` lines are removed given that the play is
already targeting the group `hosts`.