Add maas_external_hostname variable
This commit splits the `maas_external_ip_address` variable so that
different values can be configured in MaaS's `target_hostname` and
`url` settings.  This is required for environments where OpenStack
endpoints are not resolvable by the MaaS pollers, which requires
setting an IP for `target_hostname` and the internally resolvable
hostname in the `url`.

Closes: #311