LA-378 Fix swift-recon release identifier
This change modifies the swift-recon MaaS plugin to allow the directory
containing swift-recon to be specified. This change removes the need for
the plugin to try to discover the version.

Checking `/etc/openstack-release` for the version to use in general
works however due to the ordering of playbooks in leapfrog upgrades this
causes the wrong version to be used. The use of this approach also
lengthens the time of plugin failure during an upgrade given the
separation between creating the release file and updating Swift.