Set file descriptor limit for poller to 16384
When run on 14.04, the poller logs that it is running with less than
8192 as the file descriptor limit. When run on 16.04, the poller doesn't
write a log file but does run with a 16384 file descriptor limit. This
change both unifies the file descriptor limit across the board and
ensures logging is enabled on both versions.

Additionally, this change adds a check on rackspace-monitoring-poller's
file descriptor usage, creating a warning alert at 80% capacity and a
critical alert at 90% capacity of the 16384 limit.