Add support for MaaS Poller (#288)
* Add support for MaaS Poller

This change adds support for rackspace-monitoring-poller, similarly to
how rackspace-monitoring-agent is handled. The appropriate poller configs
are put into place, new private versions of lb_api_check_* templates are
added for the poller along with ping and ssh checks, and the poller is
started and restarted when necessary throughout the process on both
Ubuntu 14.04 and 16.04.

In order to have the poller installed and setup, the
maas_private_monitoring_enabled variable needs to be set to True,
and a value for maas_private_monitoring_zone must be set.

Fixes rcbops/rpc-openstack#2184

* Set file descriptor limit for poller to 16384

When run on 14.04, the poller logs that it is running with less than
8192 as the file descriptor limit. When run on 16.04, the poller doesn't
write a log file but does run with a 16384 file descriptor limit. This
change both unifies the file descriptor limit across the board and
ensures logging is enabled on both versions.

Additionally, this change adds a check on rackspace-monitoring-poller's
file descriptor usage, creating a warning alert at 80% capacity and a
critical alert at 90% capacity of the 16384 limit.