Rearrange poller restart for standalone ceph
Standalone ceph deploys are currently erroring out as
rackspace-monitoring-poller won't be deployed there, but the restart
task for the poller first checks for `groups['shared-infra_hosts']`.
That group doesn't exist in the stripped down inventory where this is
being run, so flipping the ordering to first check for the poller's
deployment will keep us from continuing on in the `when` checks for
things like that group which might not be available.