Cleaned up and reworked the layout
Fixed a copy/paste error in ceph_monitoriing.py

After discussing it with Andy we decided to put the rgw status
check on the rgw hosts themselves for now.  So I had to rework he
group placement

Cleaned up the template since we also decided to not check all the
rgw nodes at one time. Created a CRITICAL and WARNING check.