HUM-304 Allow use of other ansible bin and inv
For non OpenStack deploys we shouldnt enforce the use of
openstack-ansiblei binary.

For example, for deploying Ceph standalone, we may have a separate
inventory that isn't in openstack-ansible and a binary that isn't the
"openstack-ansible" binary.

We should allow MaaS to not force the use of openstack-ansible binary
and inventory used by ansible, and instead allow these to be set by
environment variables.