RE-360 Include Reno notes in release notes
This changes adds support for the rpc-gating generate_release_notes
hook.

New releases will now automatically have release notes generated from
those managed by Reno. If there are no Reno notes for the release that
will be explicitly stated in the release notes.

This code is ported from that used by rpc-openstack, with the use of
rpc-differ removed. The same basic structure is maintained to simplify
working on the two projects and to make it easier to unify if this ever
becomes something that is centrally managed.