Expose MaaS check details for doc automation
This change introduces a module called maas_check_util which includes a
generator method called check_details. The method yields check names and
corresponding details about the check, including the alarms included in
a check along with their alerting criteria and any variables used in
configuring said check.

The module is used by a Sphinx extension of the same name which is
enabled on the rcbops/privatecloud-docs repo to generate check
documentation from this source rather than having that done by hand.