[CEPHSTORA-22] Fix MaaS to allow Ceph Luminous support
In order to support luminous, we need to allow access to client.raxmon
for the mgr service (which is required to perform the ceph commands we
run in the script). In order for this to work from an upgrades
perspective, we need to first ensure the client.raxmon exists, and then
add the permissions - performing "get-or-create" with the new
permissions will fail of the user already exists without those
permissions.

Additionally - the health_services stanza from ceph status has been
removed.
https://www.spinics.net/lists/ceph-devel/msg37767.html
This was not a useful measurement and was removed. This patch removes
that check/metric and adjusts the check to use the "mon_in" metric which
will calculate if the mon is in and up.