Add setup_maas_tigkstack var default to False
We shouldn't need to install or setup the tigkstack grafana bits unless
we are doing upgrades. As this is causing issues with SNI/SSL on the apt
keys for influxdb we should skip installing this.

A flag is added (defaulting not defined), so that we can easily enable
this for testing downtime on upgrades, setting the
"setup_maas_tigkstack" variable to True.