Update maas_influxdata_key variable
In #414 we tried making this include conditional, however this playbook
is still being included when run w/ older versions of ansible (1.9.6
as an example).

Instead, we update maas_influxdata_key to have an id and keyserver key.
This allows apt_key to work since it doesn't need to speak to
https://repos.influxdata.com/ directly.  Using ansible's apt_key with
https://repos.influxdata.com/ fails because of SNI and:

1. https://github.com/ansible/ansible/pull/32053/ (not backported to
   2.3.2.0)
2. ansible 1.9.6 doesn't seem to support SNI on Trusty (python < 2.7.9)