RE-1296 Use rpc-ceph gating script for deployment
This change replaces the use of the deprecated script run_tests.sh with
gating/post_merge_test/run.

Both rpc-maas and rpc-ceph use RE_JOB_SCENARIO but in different ways,
the variable is overridden here to ensure Ceph is deployed correctly.

TEST_RPC_MAAS ensures that the rpc-ceph scripts do not test the MaaS
configuration because that will be done by the tests defined here.