Split warning and critical to separate alarms
This commit splits the logical warning and critical states into separate
alerts.

A story is required to understand why this is necessary.  Currently, the
process that transform MaaS events into CORE tickets null routes MaaS events
matching the WARNING status.  To work around this, both the logical warning
and critical states were mapped to the MaaS CRITICAL state to ensure a ticket
was generated in CORE.  This ensured support received a ticket and could
investigate.  However, MaaS also does not re-issue an event just because a
status message (i.e. from warning to BREACH) changed, which is a problem
when you need to distinguish between logical warning and critical states for
reporting.

So, after this change we will receive a NEW alert when a metric transitions
from a logical warning state to a logical critical state.