vars: update maas_check_timeout to 59sec
Now that timeouts are properly applied to agent.plugin checks, the
`maas_check_timeout` value should be raised from 30s to 59s (1s short of the
period). If not, I suspect that we're going to see a higher percentage of plugin
timeouts as the default value is currently half that of what we have been using
all this time (60s default in rackspace monitoring).

Fixes: TURTLES-745