JIRA: REO-218
Fix missing modules in venv

During our upgrades in phobos we found that the script wrapper
was pointing to an old venv.  After upgrading it to the new
maas-1.7.0 venv, we flooded the queue with tickets due to missing
venv modules.  I pulled all of the missing values from
/tmp/pip-constraints.txt and updated the vars/maas-agent.yml to
include them.  After re-running the site.yml all alearts cleared.