Tag maas_rally metrics with influxdb_database
This change adds an 'influxdb_database' tag to each maas_rally metric sent to
an influxdb endpoint.  This simplifies routing when the influxdb endpoint is
actually an intermediary that ignores the database specified in the write (e.g.
telegraf's http_listener input plugin).

This enables routing to different influxdb databases in telegraf like this:

```
# maas_rally sends points here
[[inputs.http_listener]]
  service_address = ":8186"

# points with no 'influxdb_database' tag go to the 'telegraf' database
# (default case)
[[outputs.influxdb]]
  urls = ["http://localhost:8086"]
  database = "telegraf"
  [outputs.influxdb.tagdrop]
    influxdb_database = ["*"]

# points with a 'influxdb_database=maas_rally' tag go to the 'maas_rally'
# database
[[outputs.influxdb]]
  urls = ["http://localhost:8086"]
  database = "maas_rally"
  tagexclude = ["influxdb_database"]
  [ouputs.influxdb.tagpass]
    influxdb_database = ["maas_rally"]
```