Revert 68ac2f9 - addition of influxdb_database tag
This change reverts 68ac2f9. There was already a mechanism for adding influxdb
tags, so rather than add it in the plugin, it's best to leave it up to the
deployer to override if necessary using ansible vars.

For example, the following in `user_maas_vars.yml` would accomplish the same
outcome:

```
maas_rally_influxdb_tags:
  environment: "{{ maas_rally_influxdb_environment_name }}"
  influxdb_database: "{{ maas_rally_influxdb_database }}"
```