Simplify maas_rally pip dependencies & constraints
This change removes maas_rally's dependency on the MaaS virtual environment's
pip requirements and constraints.  The catalyst for this was gating failures
related to the version of python-magnumclient pinned in the standard maas
virtual environment.

Troubleshooting revealed that maas_rally only needed monitorstack from the maas
requirements list.

Itemized changes:

- Remove task to copy pip-constraints.txt during maas_rally venv build
- Remove references to pip-constraints.txt from maas_rally venv build
- Remove references to maas_pip_packages ansible var from maas_rally venv build
- Add monitorstack dependency to maas_rally_pip_packages ansible var
- Swap openstacksdk upper constraint for a keystoneauth1 lower constraint (an
  improvement of 551dd6b)