Update maas_checks_util for Python 2, new checks
Per INTAKERPC-120 and TURTLES-802, this utility needs to be changed to
work on Python 2.7 as well as updated to handle newly added checks.
Additionally, there was an ask to utilize the categorization that check
templates now expose.

This change updates the Jinja undefined handler to work with both Python
2 and Python 3 as well as the necessary changes to stop ignoring some
newer checks that we were skipping.

Additionally, the module was renamed slightly so it's easier to talk
about both the docs end of this and the rpc-maas end of it.