Move check configuration docs into this repo
MaaS configuration docs used to be handled within privatecloud-docs by doing
a checkout of this repo, importing maas_checks_config out of a temporary
directory, and rendering Sphinx tables from that output. After meeting
with several docs folks it was suggested to have the docs live fully
within this repo and be served via Deconst out of here. This reduces the
complexity of building privatecloud-docs as it doesn't depend on an
external source and allows this documentation to change on its own.

This introduces per-category docs via the index, and a single-page
option linked at the top.