Remove the filebeat check
The filebeat checks are in conflict with the direction we're taking the
rest of the Elastic stack. The Elastic stack work is now running running
beats (not just filebeat) on hosts machines and is limiting beat
installations within containers. This is on top of the fact that the
RAX-MaaS checks for filebeat are largely meaningless and causes cascading
alerts/tickets especially when any one process is disabled or otherwise
removed. Log shipping to the logging host is taken care of with rsyslog,
and journald (journal-remote). While filebeat is useful, it's only an
analytic tool and not something that is mission critical requiring an
alert if it stops for arbitrary reasons.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>