Remove the elasticsearch maas check
The elasticsearch maas check is not doing a whole lot and is more noise
than value. This is on top of the fact that the elasticsearch checks are
not made for a deployment where elasticsearch/logstash are deployed in a
high performance cluster.

This change follows on with the removal of the filebeat check.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>