Support category/check/alarm doc descriptions
To support the need for prose documentation alongside the automated
check docs that are rendered, this change introduces a `stubs` directory
that includes documentation stubs that are inserted during the build
at the appropriate spot for said stub. Each category and check includes
a "title.rst" stub file for category and check-wide descriptions to go
under their respective titles, and alarm descriptions are named after
the alarm.

This includes empty stubs for everything right not that makes a complete
build. If stubs are missing for any category, check, or alarm, that will
cause the build to emit a warning log, which at current will kill the
build. If stubs are empty, which all of them are right now, the build
will emit an info message to the log stating that it found no content.