Remove tox for Ceph jobs
The RPC-Ceph gate already has a funcitonal test that runs and deploys
MaaS. Utilizing tox from the MaaS side means these tests don't work the
same way and causes some issues. We can remove the tox portion and just
utilize the rpc-ceph job directly, which will then handle the
configuration and deployment of MaaS from the rpc-maas repo that was
cloned as part of the test run.