Add bandit to tox.ini
This change adds support for the bandit security static analyzer tool,
per https://rpc-openstack.atlassian.net/browse/TURTLES-1017.

This is currently configured to run bandit against any Python code it
finds inside the playbooks directory, which is all of the plugin code
that the maas checks run. It is additionally configured to ignore two
low severity checks that are about the existence of import statements
for subprocess and lxml. We use those libraries intentionally, so we
don't need to be warned that their introduction to our codebase are
potentially vulnerable to issues. Ignoring those two warnings doesn't
ignore warnings emitted around specific vulnerabilities found.