Run bandit twice for extra logging
Bandit currently is not able to output issues at severities lower than
what it's configured to run and evaluate pass/fail status of, but we
still need to know about those lower severity issues rather than ignore
them. Since bandit runs in a matter of seconds, this change runs bandit
at the lowest level to get a log of every possible issue, but it ignores
the exit code (the leading '-' does this). The second run of bandit runs
at the most strict level, and that's what we pass or fail based on.