Update maas to support recent version of OpenStack
These changes allow maas to run with modern openstack APIs. This work
was done as part of the tools team deployment within an RPCO lab
environment where maas was deployed on the Rocky version of OpenStack on
ubuntu 18.04. While the deployment validation was done on Rocky the
changes will benifit the tooling as a whole allowing us to better serve
our existing customers.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>