Re-introduce the run-tests script
This change documents how to run tests locally. Which should help
developers and operators setup test environments.

To fix the gate several playbook changes have been allowing the
playbooks to converge when running local testing.

The inclusion of the tigstack tests have been removed for the following
reasons:
1 - its broken
2 - its not deployed anywhere

The test playbooks maas-raxmon* have been moved into the standard
playbook directory. This was done because these playbooks require a
library that is not available from within the tests directory path.
Without the ability to access the library these playbooks are broken. By
moving the playbooks into the standard playbook directory the
functionality these playbooks provide was restored.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>