Remove unused / maintained monitoring solutions and options (#612)
* Removing humming bird monitoring

The humming bird project has been abandoned and will no longer be
deployed or supported anywhere. This change removes all of that dead
weight.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* Remove the tigk stack

None of this has ever been used. We're deleting it because its all just
dead weight.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* remove site-osp

This playbook is not used.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* variable and conditional consolodation

We have a lot of vars being repeated, in a lot of different files. This
change pulls what we can, simply, together. By doing this we can also
simplify some of the conditions we use to run common-tasks and options.

Variables have also been normalized where possible

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

* Fix deprecation warnings

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>