Bump Ansible version to 2.7.11
The most notable change upgrading Ansible from 2.6.12 to 2.7.11 is that
Ansible 2.7+ requires setuptools in order to build an isolated
virtualenv with --no-system-site-packages (default). Our typical OSA
deployments alternatively replace the --index-url with that of the repo
container for all pip related tasks. As setuptools is now required, but
not part of the repo, it will fail. These changes will temporarily
remove the index-url from the pip.conf to successfully build the
virtualenv.

Signed-off-by: Nathan Pawelek <nathan.pawelek@rackspace.com>