Initial content commit for documentation overhaul
This change is the overarching commit for rpc-maas stub documentation
updates. Changes to the repository as a result of dynamic variable
generation facilitated modifications to the original
maas_checks_config.py. This ensured functionality with the latest
repository content. The plugin walks check templates and creates stubs
for all checks, alarms, and their associated thresholds.

Additional changes include:

 * Removed checks and alarms that no longer exist in the repository.
   Mostly Kubernetes, non-existent, and non-value add alarms.
 * Copy Sphinx template settings and static files from the generic RPC
   documentation.
 * Various minor template changes.

Signed-off-by: Nathan Pawelek <nathan.pawelek@rackspace.com>