Add functionality for verifying CPU softnet stats
If a softnet backlog is full, it will silently drop packets without any
logged indication of it happening. The second column of
/proc/net/softnet_stats will increment on a per-CPU basis. This
functionality simply sums these values while ignoring what CPU the drop
happened on. The threshold has been configured as 0 because any value
for these indicates that the kernel must be tuned accordingly.

Closes #677

Signed-off-by: Nathan Pawelek <nathan.pawelek@rackspace.com>