Update hp_monitoring.py to catch predictive failures
This fixes a bug with the hp monitoring plugin which was only detecting
failures associated with logical volumes. When a physical drive is marked
as 'Predictive Failure' with the ssacli command line utility, the status
does not cause the logical volume to degrade. This change separates the
logicaldrive and physical drive metrics to ensure that predictive
failures are also accounted for.

Signed-off-by: Nathan Pawelek <nathan.pawelek@rackspace.com>