JIRA: RPCOS-33 Fix ceph_health for luminous
When installing rpc-maas onto OSP13 using ceph, the ceph_health
check was failing.  It looks like this was due to a breaking
change in PR 702 assuming that either 'overall_status' or
'status' checks exist exclusively. Luminous has both with
'status' being set to 'HEALTH_OK' and 'overall_status' set
to 'HEALTH_WARN' to let users know not to use it.