JIRA RPCOS-44 Add metrics for health checks
The ceph health output provides a 'checks' dictionary for when the cluster
status is something other than 'HEALTH_OK'. Before we can start setting
up checks and alerts for these, we need them converted into metrics for
the rackspace monitoring agent.

These changes create a different 'health_checks' subcommand that will
default to 'HEALTH_OK(2)' for each and change it to the status value
in the 'checks' dictionary if found. I had to break these into sections
as each check can have a max of 50 metrics.  The '--section' argument
allows for mon,mgr,osds,device_health,data_health or misc. These line
up with the sections given in the following ceph document:

https://docs.ceph.com/docs/master/rados/operations/health-checks/

I also fixed a couple pep8 spacing issues in the process.