Vastly speed up bond detection
Previously, each interface in ansible_interfaces was being iterated to
find interfaces of type 'bonding'. On controllers hosting a large number
of neutron networks, the number of interfaces can be in the thousands.

This made the bond interface task take an very long time, especially
coupled with a large number of hostvars, since hostvars were being
loaded thousands of times.

This patch instead simply looks for the `ansible_bond0` and
`ansible_bond1` facts and adds the bond if the fact exists. This means
hostvars are loaded only twice per host instead number of interfaces
times.